# move: source service eagerly hydrates modules in memory